+++ docs/applications.md
@@ -76,3 +76,14 @@ exceptions that occur within the application:
 * `app.add_exception_handler(exc_class_or_status_code, handler)` - Add an error handler. The handler function may be either a coroutine or a regular function, with a signature like `func(request, exc) -> response`.
 * `@app.exception_handler(exc_class_or_status_code)` - Add an error handler, decorator style.
 * `app.debug` - Enable or disable error tracebacks in the browser.
+
+### Storing state on the app instance
+
+You can store arbitrary extra state on the application instance, using the
+generic `app.state` attribute.
+
+For example:
+
+```python
+app.state.ADMIN_EMAIL = 'admin@example.org'
+```
