+++ starlette/middleware/errors.py
@@ -1,5 +1,5 @@
-import html
 import asyncio
+import html
 import inspect
 import traceback
 import typing
